+++ example/lib/main.dart
@@ -9,9 +9,7 @@ class MyApp extends StatelessWidget {
   Widget build(BuildContext context) {
     return MaterialApp(
       title: 'Blur Dialog',
-      theme: ThemeData(
-        primarySwatch: Colors.blue,
-      ),
+      theme: ThemeData(primarySwatch: Colors.blue),
       home: MyHomePage(title: 'Example BlurDialog'),
     );
   }
@@ -30,320 +28,231 @@ class MyHomePage extends StatelessWidget {
       body: ListView(
         padding: EdgeInsets.all(20),
         children: <Widget>[
-          OutlinedButton(
-            child: Text("Show CustomDialog from extension"),
-            onPressed: () {
-              Future.delayed(Duration(seconds: 3)).showProgressDialog(context,
-                  title: Text("This is just an ordinary loading"),
-                  message:
-                      Text("This is fastest and coolest way to show a dialog"));
-            },
-          ),
-          OutlinedButton(
-            child: Text("NAlertDialog show"),
-            onPressed: () {
-              NAlertDialog(
-                title: Text("Test"),
-                content: Text("Iya iya"),
-                blur: 2,
-              ).show(context, transitionType: DialogTransitionType.Bubble);
-            },
-          ),
-          OutlinedButton(
-            child: Text("NDialog can do it too!"),
-            onPressed: () {
-              NDialog(
-                title: Text("Test"),
-                dialogStyle: DialogStyle(),
-                content: Text("Iya iya"),
-              ).show(context);
-            },
-          ),
-          OutlinedButton(
-            child: Text("Progress Dialog"),
-            onPressed: () async {
-              ProgressDialog progressDialog = ProgressDialog(
-                context,
-                blur: 0,
-                dialogTransitionType: DialogTransitionType.Shrink,
-                title: Text("Title"),
-                message: Text("Message"),
-                onDismiss: () {
-                  print("Do something onDismiss");
-                },
-              );
-              progressDialog.setLoadingWidget(CircularProgressIndicator(
-                valueColor: AlwaysStoppedAnimation(Colors.red),
-              ));
-              progressDialog.setMessage(
-                  Text("Please Wait, Injecting your phone with my virus"));
-              progressDialog.setTitle(Text("Loading"));
-              progressDialog.show();
+          ///Show NAlertDialog
+          _button("Show NAlertDialog", () {
+            NAlertDialog(
+              dialogStyle: DialogStyle(titleDivider: true),
+              title: Text("NAlertDialog"),
+              content: Text("This is NAlertDialog's content"),
+              actions: <Widget>[
+                TextButton(
+                    child: Text("Okay"),
+                    onPressed: () => Navigator.pop(context)),
+                TextButton(
+                    child: Text("Close"),
+                    onPressed: () => Navigator.pop(context)),
+              ],
+            ).show(context);
+          }),
 
-              await Future.delayed(Duration(seconds: 5));
+          ///Show NDialog
+          _button("Show NDialog", () {
+            NDialog(
+              dialogStyle: DialogStyle(titleDivider: true),
+              title: Text("NDialog"),
+              content: Text("This is NDialog's content"),
+              actions: <Widget>[
+                TextButton(
+                    child: Text("Okay"),
+                    onPressed: () => Navigator.pop(context)),
+                TextButton(
+                    child: Text("Close"),
+                    onPressed: () => Navigator.pop(context)),
+              ],
+            ).show(context);
+          }),
 
-              progressDialog.setMessage(Text("I mean, virus of love :*"));
-              progressDialog.setLoadingWidget(null);
-              progressDialog.setTitle(Text("Just Kidding"));
+          ///Show ProgressDialog traditional way
+          _button("Show ProgressDialog traditional way", () async {
+            ProgressDialog progressDialog = ProgressDialog(
+              context,
+              blur: 10,
+              title: Text("Title of ProgressDialog"),
+              message: Text("Content of ProgressDialog"),
+              onDismiss: () => print("Do something onDismiss"),
+            );
+            progressDialog.show();
 
-              await Future.delayed(Duration(seconds: 5));
+            //Set loading with red circular progress indicator and change the title and message
+            await Future.delayed(Duration(seconds: 5));
+            progressDialog.setLoadingWidget(CircularProgressIndicator(
+                valueColor: AlwaysStoppedAnimation(Colors.red)));
+            progressDialog.setTitle(Text("ProgressDialog Title changed"));
+            progressDialog
+                .setMessage(Text("It will change again after 5 seconds"));
 
-              progressDialog.dismiss();
-            },
-          ),
-          OutlinedButton(
-            onPressed: () async {
-              CustomProgressDialog progressDialog = CustomProgressDialog(
-                context,
-                blur: 10,
-                onDismiss: () {
-                  print("Do something onDismiss");
-                },
-              );
-              progressDialog.setLoadingWidget(CircularProgressIndicator(
-                valueColor: AlwaysStoppedAnimation(Colors.red),
-              ));
-              progressDialog.show();
+            //Show dialog
+            progressDialog.show();
 
-              await Future.delayed(Duration(seconds: 5));
-              progressDialog.setLoadingWidget(null);
+            //After 5 seconds, change the title, message and loading widget to null (default widget is CircularProgressIndicator)
+            await Future.delayed(Duration(seconds: 5));
+            progressDialog.setLoadingWidget(null);
+            progressDialog.setMessage(
+                Text("ProgressDialog will dismiss after 5 seconds"));
+            progressDialog.setTitle(Text("Last changes"));
 
-              await Future.delayed(Duration(seconds: 5));
+            //After 5 seconds, dismiss the dialog
+            await Future.delayed(Duration(seconds: 5));
+            progressDialog.dismiss();
+          }),
 
-              progressDialog.dismiss();
-            },
-            child: Text("Custom Progress Dialog"),
-          ),
-          OutlinedButton(
-            onPressed: () async {
-              print(await ProgressDialog.future(
-                context,
-                blur: 0.0,
-                future: Future.delayed(Duration(seconds: 5), () {
-                  return "HIYAAA";
-                }),
-                onProgressError: (error) {
-                  print("Do something onProgressError");
-                },
-                onProgressFinish: (data) {
-                  print("Do something onProgressFinish");
-                },
-                onDismiss: () {
-                  print("Dismissed");
-                },
-                message: Text("Please Wait"),
-                cancelText: Text("Batal"),
-                title: Text("Loging in"),
-              ));
-            },
-            child: Text("Progress Dialog Future"),
-          ),
-          OutlinedButton(
-            child: Text("Custom Progress Dialog Future"),
-            onPressed: () async {
-              print(await CustomProgressDialog.future(
-                context,
-                backgroundColor: Colors.blue.withOpacity(.5),
-                future: Future.delayed(Duration(seconds: 5), () {
-                  return "WOHOOO";
-                }),
-                onProgressError: (error) {
-                  print("Do something onProgressError");
-                },
-                onProgressFinish: (data) {
-                  print("Do something onProgressFinish");
-                },
-                onDismiss: () {
-                  print("Dismissed");
-                },
-              ));
-            },
-          ),
-          OutlinedButton(
-            onPressed: () async {
-              NAlertDialog(
-                dialogStyle: DialogStyle(titleDivider: true),
-                title: Text("Hi, This is NAlertDialog"),
-                content: Text("And here is your content, hoho... "),
+          ///Show ProgressDialog traditional way
+          _button("Show CustomProgressDialog traditional way", () async {
+            CustomProgressDialog progressDialog = CustomProgressDialog(
+              context,
+              blur: 10,
+              onDismiss: () => print("Do something onDismiss"),
+            );
+
+            //Set loading with red circular progress indicator
+            progressDialog.setLoadingWidget(CircularProgressIndicator(
+                valueColor: AlwaysStoppedAnimation(Colors.red)));
+
+            //Show dialog
+            progressDialog.show();
+
+            //After 5 seconds, change the loading widget to null (default widget is CircularProgressIndicator)
+            await Future.delayed(Duration(seconds: 5));
+            progressDialog.setLoadingWidget(null);
+
+            //After 5 seconds, dismiss the dialog
+            await Future.delayed(Duration(seconds: 5));
+            progressDialog.dismiss();
+          }),
+
+          ///Show ProgressDialog with Future's static function
+          _button("Show ProgressDialog.future()", () {
+            ProgressDialog.future(
+              context,
+              title: Text("ProgressDialog.future()"),
+              message: Text("This will finish after 5 seconds"),
+              future: Future.delayed(
+                  Duration(seconds: 5), () => "The end of future (5 seconds)"),
+              onProgressError: (error) => print("Do something onProgressError"),
+              onProgressFinish: (data) =>
+                  print("Do something onProgressFinish"),
+              onDismiss: () => print("Dismissed"),
+            ).then((value) => print(value));
+          }),
+
+          ///Show CustomProgressDialog with Future's static function
+          _button("Show CustomProgressDialog.future()", () {
+            CustomProgressDialog.future(
+              context,
+              future: Future.delayed(
+                  Duration(seconds: 5), () => "The end of future (5 seconds)"),
+              onProgressError: (error) => print("Do something onProgressError"),
+              onProgressFinish: (data) =>
+                  print("Do something onProgressFinish"),
+              onDismiss: () => print("Dismissed"),
+            ).then((value) => print(value));
+          }),
+
+          ///Show Progress Dialog from NDialog Future's extensions
+          _button("Show ProgressDialog from extension", () async {
+            Future.delayed(Duration(seconds: 3)).showProgressDialog(
+              context,
+              title: Text("Title of ProgressDialog"),
+              message: Text("This dialog will be close after 3 seconds"),
+            );
+          }),
+
+          ///Show Custom Progress Dialog from NDialog Future's extensions
+          _button("Show CustomProgressDialog from extension", () async {
+            Future.delayed(Duration(seconds: 3))
+                .showCustomProgressDialog(context);
+          }),
+
+          ///Show NAlertDialog with blur background
+          _button("NAlertDialog with blur background", () {
+            DialogBackground(
+              blur: 5,
+              dialog: AlertDialog(
+                title: Text("NAlert Dialog"),
+                content: Text("NAlertDialog with Blur background"),
                 actions: <Widget>[
                   TextButton(
-                    child: Text("You"),
-                    onPressed: () {},
-                  ),
+                      child: Text("Okay"),
+                      onPressed: () => Navigator.pop(context)),
                   TextButton(
-                    child: Text("Are"),
-                    onPressed: () {},
-                  ),
-                  TextButton(
-                    child: Text("Awesome"),
-                    onPressed: () {},
-                  )
+                      child: Text("Close"),
+                      onPressed: () => Navigator.pop(context)),
                 ],
-              ).show(context);
-            },
-            child: Text("Show NAlertDialog"),
-          ),
-          OutlinedButton(
-            onPressed: () async {
-              NDialog(
-                dialogStyle: DialogStyle(titleDivider: true),
-                title: Text("Hi, This is NDialog"),
-                content: Text("And here is your content, hoho... "),
-                actions: <Widget>[
-                  TextButton(
-                    child: Text("You"),
-                    onPressed: () {},
-                  ),
-                  TextButton(
-                    child: Text("Are"),
-                    onPressed: () {},
-                  ),
-                  TextButton(
-                    child: Text("Awesome"),
-                    onPressed: () {},
-                  )
-                ],
-              ).show(
-                context,
-              );
-            },
-            child: Text("Show NDialog"),
-          ),
-          OutlinedButton(
-            child: Text(
-              "Show Alert Dialog With Blur Background",
-              textAlign: TextAlign.center,
-            ),
-            onPressed: () async {
-              DialogBackground(
-                barrierColor: Colors.red.withOpacity(.5),
-                blur: 0,
-                dialog: AlertDialog(
-                  title: Text("Alert Dialog"),
-                  content: Text(
-                      "Wohoo.. This is ordinary AlertDialog with Blur background"),
-                  actions: <Widget>[
-                    TextButton(
-                      child: Text("You"),
-                      onPressed: () {},
-                    ),
-                    TextButton(
-                      child: Text("Are"),
-                      onPressed: () {},
-                    ),
-                    TextButton(
-                      child: Text("Awesome"),
-                      onPressed: () {},
-                    )
-                  ],
-                ),
-              ).show(
-                context,
-              );
-            },
-          ),
-          OutlinedButton(
-            child: Text(
-              "Show Alert Dialog With Custom Background Color",
-              textAlign: TextAlign.center,
-            ),
-            onPressed: () async {
-              NAlertDialog(
-                backgroundColor: Colors.red.withOpacity(.5),
-                blur: 0,
-                title: Text("Alert Dialog"),
-                content: Text(
-                    "Wohoo.. This is ordinary AlertDialog with Custom Color background"),
-                actions: <Widget>[
-                  TextButton(
-                    child: Text("You"),
-                    onPressed: () {},
-                  ),
-                  TextButton(
-                    child: Text("Are"),
-                    onPressed: () {},
-                  ),
-                  TextButton(
-                    child: Text("Awesome"),
-                    onPressed: () {},
-                  )
-                ],
-              ).show(context);
-            },
-          ),
-          OutlinedButton(
-            child: Text(
-              "Dialog Extension",
-              textAlign: TextAlign.center,
-            ),
-            onPressed: () async {
-              await AlertDialog(
-                title: Text("Alert Dialog"),
-                content: Text(
-                    "Wohoo.. This is ordinary AlertDialog with Blur background"),
-                actions: <Widget>[
-                  TextButton(
-                    child: Text("You"),
-                    onPressed: () {},
-                  ),
-                  TextButton(
-                    child: Text("Are"),
-                    onPressed: () {},
-                  ),
-                  TextButton(
-                    child: Text("Awesome"),
-                    onPressed: () {},
-                  )
-                ],
-              ).show(context);
-            },
-          ),
-          OutlinedButton(
-            child:
-                Text("CupertinoDialog Extension", textAlign: TextAlign.center),
-            onPressed: () async {
-              await CupertinoAlertDialog(
-                title: Text("Alert Dialog"),
-                content: Text(
-                    "Wohoo.. This is ordinary AlertDialog with Blur background"),
-                actions: <Widget>[
-                  TextButton(
-                    child: Text("You"),
-                    onPressed: () {},
-                  ),
-                  TextButton(
-                    child: Text("Are"),
-                    onPressed: () {},
-                  ),
-                  TextButton(
-                    child: Text("Awesome"),
-                    onPressed: () {},
-                  )
-                ],
-              ).show(context);
-            },
-          ),
-          OutlinedButton(
-            child: Text("ZoomDIALOG", textAlign: TextAlign.center),
-            onPressed: () async {
-              await ZoomDialog(
-                zoomScale: 5,
-                blur: 10,
-                child: Container(
-                  child: Text("Zoom me!"),
-                  color: Colors.white,
-                  padding: EdgeInsets.all(20),
-                ),
-              ).show(context);
-            },
-          ),
+              ),
+            ).show(context);
+          }),
+
+          ///Show NAlertDialog with custom background color
+          _button("NAlertDialog with backgroundColor", () {
+            NAlertDialog(
+              backgroundColor: Colors.red.withOpacity(.5),
+              blur: 0,
+              title: Text("NAlert Dialog"),
+              content: Text("NAlertDialog with custom Color background"),
+              actions: <Widget>[
+                TextButton(
+                    child: Text("Okay"),
+                    onPressed: () => Navigator.pop(context)),
+                TextButton(
+                    child: Text("Close"),
+                    onPressed: () => Navigator.pop(context)),
+              ],
+            ).show(context);
+          }),
+
+          ///Show material alret dialog with NDialog show Extensions
+          _button("AlertDialog extensions", () {
+            AlertDialog(
+              title: Text("Material AlertDialog"),
+              content: Text("This dialog shown from NDialog extensions"),
+              actions: <Widget>[
+                TextButton(
+                    child: Text("Okay"),
+                    onPressed: () => Navigator.pop(context)),
+                TextButton(
+                    child: Text("Close"),
+                    onPressed: () => Navigator.pop(context)),
+              ],
+            ).show(context);
+          }),
+
+          ///Show cupertino alret dialog with NDialog show Extensions
+          _button("CupertinoAlertDialog extensions", () {
+            CupertinoAlertDialog(
+              title: Text("Cupertino Alert Dialog"),
+              content: Text("This dialog shown from NDialog extensions"),
+              actions: <Widget>[
+                CupertinoButton(
+                    child: Text("Okay"),
+                    onPressed: () => Navigator.pop(context)),
+                CupertinoButton(
+                    child: Text("Close"),
+                    onPressed: () => Navigator.pop(context)),
+              ],
+            ).show(context);
+          }),
+
+          ///Show zoomable widget in dialog
+          _button("Show ZoomDialog", () {
+            ZoomDialog(
+              zoomScale: 5,
+              blur: 10,
+              child: Container(
+                child: Text("Zoom me!"),
+                color: Colors.white,
+                padding: EdgeInsets.all(20),
+              ),
+            ).show(context);
+          }),
         ]
-            .map((e) => Container(
-                  margin: EdgeInsets.symmetric(vertical: 5),
-                  child: e,
-                ))
+            .map((e) =>
+                Container(margin: EdgeInsets.symmetric(vertical: 5), child: e))
             .toList(),
       ),
     );
+  }
+
+  Widget _button(String title, Function() onPressed) {
+    return OutlinedButton(child: Text(title), onPressed: onPressed);
   }
 }